requirements.txt
# Base tools
wheel
setuptools
python-xlib==0.33

# GUI packages
PyQt6>=6.4.2,<6.5.0  # Updated version constraint
PyQt6-Qt6>=6.4.2,<6.5.0  # Updated version constraint
PyQt6_sip>=13.4.1,<14.0.0
pyqt6-plugins>=6.4.2.2.2,<6.4.2.3  # Added required plugin package

# Database
aiomysql==0.2.0
aiosqlite==0.20.0
PyMySQL==1.1.1
mysql-connector-python==8.0.33  # Добавлен mysql-connector-python

# Development tools
black==24.1.1
pylint==3.0.3
pytest==8.0.0
isort==5.13.2
docformatter==1.7.5
pydocstyle==6.3.0
flake8==7.0.0
mypy==1.8.0

# Testing tools
coverage==7.6.8
pytest-cov==6.0.0

# Other dependencies
astroid>=3.0.1,<=3.1.0
async==0.6.2
asyncio==3.4.3
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dill==0.3.9
EbookLib==0.18
et_xmlfile==2.0.0
flask
idna==3.10
iniconfig==2.0.0
isort==5.13.2
lxml==5.3.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy>=1.24.0
openpyxl==3.1.5
packaging==24.1
pandas>=2.0.0
pathspec==0.12.1
pep8==1.7.1
pillow>=9.5.0
platformdirs==4.3.6
pluggy==1.5.0
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
qasync==0.27.1
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.6
Telethon==1.37.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
XlsxWriter==3.2.0

# Flask and extensions
flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5

# System dependencies
# Please install the following package using your system's package manager:
# sudo apt install -y default-mysql-client